Рефакторинг частей кода.

+++ Objective-C_level_1/DZ7/DZ7/Employee.h
@@ -10,10 +10,10 @@
 
 @interface Employee : NSObject <NSCoding>
 
-@property (nonatomic, readwrite) unsigned salary;
+@property (nonatomic, readwrite) NSUInteger salary;
 @property (nonatomic, readwrite, assign) NSString *name;
 @property (nonatomic, readwrite, assign) NSString *surname;
 
-- (instancetype)initWhithValues: (unsigned) slary name: (NSString *) name surname: (NSString *) surname;
+- (instancetype)initWhithValues: (NSUInteger) slary name: (NSString *) name surname: (NSString *) surname;
 
 @end
